Update Get-PwnedPassword to use K-anonymity only
Troy announced that direct hash and password searches will be deprecated.

https://www.troyhunt.com/enhancing-pwned-passwords-privacy-by-exclusively-supporting-anonymity/

This is great but would break the existing API methods.  I updated the script to use range checking.  In line with Troy's changes, I recommend removing the -password option and requiring securestring or sha1 only, but defer that to the owner.